restructure module hierarchy with primary and secondary ports
