+++ libs/mimir2/src/domain/mod.rs
@@ -1,3 +1,2 @@
 pub mod model;
 pub mod ports;
-pub mod usecases;
