+++ libs/mimir2/src/domain/ports/mod.rs
@@ -1,6 +1,2 @@
-pub mod explain;
-pub mod import;
-pub mod list;
-pub mod remote;
-pub mod search;
-pub mod storage;
+pub mod primary;
+pub mod secondary;
