+++ libs/mimir2/src/domain/ports/primary/search_documents.rs
@@ -1,5 +1,5 @@
 use crate::domain::model::query::Query;
-use crate::domain::ports::search::{Error, Parameters, Search};
+use crate::domain::ports::secondary::search::{Error, Parameters, Search};
 use serde::de::DeserializeOwned;
 
 pub async fn search_documents<B, D>(
