+++ libs/mimir2/src/domain/ports/secondary/search.rs
